added static swappable ObjectFactory implementation
