+++ src/CommandLine/Parser.cs
@@ -88,7 +88,7 @@ namespace CommandLine
             if (args == null) throw new ArgumentNullException("args");
 
             var factory = typeof(T).IsMutable()
-                ? Maybe.Just<Func<T>>(CreateInstance<T>)
+                ? Maybe.Just<Func<T>>(ParserSettings.ObjectFactory.Resolve<T>)
                 : Maybe.Nothing<Func<T>>();
 
             return MakeParserResult(
@@ -103,16 +103,6 @@ namespace CommandLine
                 settings);
         }
 
-        /// <summary>
-        /// Create Instance to use for binding 
-        /// Override to add container resolution 
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <returns></returns>
-        protected virtual T CreateInstance<T>()
-        {
-            return settings.DefaultObjectFactory.CreateInstance<T>();
-        }
 
         /// <summary>
         /// Parses a string array of command line arguments constructing values in an instance of type <typeparamref name="T"/>.
